feat: separate attachments view
